amounts with float. ars added.

+++ src/main/java/com/gavilan/currency/domain/Bank.java
@@ -4,17 +4,17 @@ import java.util.Hashtable;
 
 public class Bank {
 
-    private final Hashtable<Pair, Integer> rates = new Hashtable<>();
+    private final Hashtable<Pair, Float> rates = new Hashtable<>();
 
     public Money reduce(Expression source, String to) {
         return source.reduce(this, to);
     }
 
-    public void addRate(String from, String to, int rate) {
+    public void addRate(String from, String to, float rate) {
         rates.put(new Pair(from, to), rate);
     }
 
-    public int rate(String from, String to) {
+    public float rate(String from, String to) {
         if (from.equals(to)) return 1;
         return rates.get(new Pair(from, to));
     }
